+++ CoreGraphFramework/src/au/gov/asd/tac/constellation/graph/value/operations/Difference.java
@@ -25,7 +25,9 @@ import au.gov.asd.tac.constellation.graph.value.Operators;
  */
 public class Difference {
 
-    public static final String NAME = "DIFFERENCE";
+    // Explicitly calling the constructor on a String appears to be necessary to
+    // allow the registry to find the operator.
+    public static final String NAME = new String("DIFFERENCE");
 
     private Difference() {
         // added private constructor to hide implicit public constructor - S1118.